Fix dllimport/export mismatch warning.

+++ smblib/smbadd.c
@@ -2,7 +2,7 @@
 
 /* Synchronet message base (SMB) high-level "add message" function */
 
-/* $Id: smbadd.c,v 1.23 2011/10/16 09:53:10 rswindell Exp $ */
+/* $Id: smbadd.c,v 1.22 2009/03/24 20:39:35 rswindell Exp $ */
 
 /****************************************************************************
  * @format.tab-size 4		(Plain Text/Source Code File Header)			*
@@ -243,7 +243,7 @@ int SMBCALL smb_addmsg(smb_t* smb, smbmsg_t* msg, int storage, long dupechk_hash
 		}
 
 		if(msg->hdr.when_imported.time==0) {
-			msg->hdr.when_imported.time=(uint32_t)time(NULL);
+			msg->hdr.when_imported.time=time(NULL);
 			msg->hdr.when_imported.zone=0;	/* how do we detect system TZ? */
 		}
 		if(msg->hdr.when_written.time==0)	/* Uninitialized */
